+++ packages/ui5-application-writer/src/options.ts
@@ -124,3 +124,27 @@ export async function applyOptionalFeatures(
         }
     }
 }
+
+/**
+ * Generates the resource URL based on the project type and ui5 framework details.
+ *
+ * @param {boolean} isEdmxProjectType Indicates if the project type is Edmx or CAP.
+ * @param {string} [frameworkUrl] URL of the ui5 framework.
+ * @param {string} [version] version of the ui5 framework.
+ * @returns {string} - The constructed resource URL based on project type.
+ */
+export function getTemplateOptions(isEdmxProjectType: boolean, frameworkUrl?: string, version?: string): string {
+    const resourcePath = 'resources/sap-ui-core.js';
+    if (isEdmxProjectType || !frameworkUrl) {
+        // Use relative path for Edmx projects or if frameworkUrl is not available
+        return resourcePath;
+    } else {
+        // return the full URL for CAP projects
+        let url = frameworkUrl;
+        if (version) {
+            url += `/${version}`;
+        }
+        url += `/${resourcePath}`;
+        return url;
+    }
+}