+++ packages/fiori-elements-writer/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @sap-ux/fiori-elements-writer
 
+## 0.23.2
+
+### Patch Changes
+
+-   Updated dependencies [22e4ad8]
+    -   @sap-ux/ui5-config@0.23.1
+    -   @sap-ux/odata-service-writer@0.21.1
+    -   @sap-ux/ui5-application-writer@0.27.2
+    -   @sap-ux/fe-fpm-writer@0.25.0
+    -   @sap-ux/ui5-test-writer@0.4.0
+
+## 0.23.1
+
+### Patch Changes
+
+-   8b7ee1f: Do not add reuse libraries to the ui5-local.yaml
+-   Updated dependencies [8b7ee1f]
+    -   @sap-ux/ui5-application-writer@0.27.1
+
+## 0.23.0
+
+### Minor Changes
+
+-   c2359077: [BREAKING CHANGE] Change TypeScript transpile target to ES2021 to align with NodeJS 18+
+
+### Patch Changes
+
+-   Updated dependencies [c2359077]
+    -   @sap-ux/fe-fpm-writer@0.25.0
+    -   @sap-ux/odata-service-writer@0.21.0
+    -   @sap-ux/ui5-application-writer@0.27.0
+    -   @sap-ux/ui5-config@0.23.0
+    -   @sap-ux/ui5-test-writer@0.4.0
+
 ## 0.22.0
 
 ### Minor Changes