+++ packages/ui5-proxy-middleware/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @sap-ux/ui5-proxy-middleware
 
+## 1.4.1
+
+### Patch Changes
+
+-   Updated dependencies [22e4ad8]
+    -   @sap-ux/ui5-config@0.23.1
+
+## 1.4.0
+
+### Minor Changes
+
+-   c2359077: [BREAKING CHANGE] Change TypeScript transpile target to ES2021 to align with NodeJS 18+
+
+### Patch Changes
+
+-   Updated dependencies [c2359077]
+    -   @sap-ux/logger@0.6.0
+    -   @sap-ux/ui5-config@0.23.0
+
 ## 1.3.16
 
 ### Patch Changes